Convert netbackup to use pflag.
- The change to pflag allows for cleaner use of multiple string
  config files (upcoming change).
- Also removed flag.go and moved the flag parsing routines to main.go.
