+++ go.mod
@@ -5,4 +5,5 @@ go 1.13
 require (
 	github.com/BurntSushi/toml v0.3.1
 	github.com/marcopaganini/logger v0.1.2
+	github.com/spf13/pflag v1.0.5 // indirect
 )
